Feature/#24 (#31): 모든 유저 채팅 가져오는 기능 수정
* [refactor] sse 테스트코드 전환

* [refactor] ChatMessageResponse 변환으로 인한 수정

* [refactor] AI Server Request 형식 수정

* [refactor] weekweather api 기능 변경

- 기존: 요청이 들어올때마다 open api 호출 후 전처리
- 변경: 스케줄링을 이용해 DB에 매일 업데이트
만약, 데이터가 없을시 요청시점에 DB에 업데이트

* [bug-fix] successHandler 주석 처리 해제

* [bug-fix] 모든 유저 채팅 가져오는 기능 수정
