+++ src/main/java/org/ktb/chatbotbe/domain/chat/controller/ChatController.java
@@ -4,6 +4,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.ktb.chatbotbe.domain.chat.dto.controller.request.ChatMessageCreateRequest;
 import org.ktb.chatbotbe.domain.chat.dto.service.response.*;
+import org.ktb.chatbotbe.domain.chat.dto.service.response.strategy.ChatMessageResponse;
+import org.ktb.chatbotbe.domain.chat.dto.service.response.strategy.DoneResponse;
+import org.ktb.chatbotbe.domain.chat.dto.service.response.strategy.NewChatResponse;
+import org.ktb.chatbotbe.domain.chat.dto.service.response.strategy.TitleAIResponse;
 import org.ktb.chatbotbe.domain.chat.service.ChatService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.codec.ServerSentEvent;
