--- src/main/java/org/ktb/chatbotbe/domain/chat/dto/service/response/ChatAIResponse.java
@@ -1,5 +1,0 @@
-package org.ktb.chatbotbe.domain.chat.dto.service.response;
-
-
-public interface ChatAIResponse {
-}
