--- src/main/java/org/ktb/chatbotbe/domain/chat/dto/service/response/DoneResponse.java
@@ -1,8 +1,0 @@
-package org.ktb.chatbotbe.domain.chat.dto.service.response;
-
-import lombok.Getter;
-
-@Getter
-public class DoneResponse implements ChatAIResponse{
-    private final ChatMessageType type = ChatMessageType.DONE;
-}
