+++ src/main/java/org/ktb/chatbotbe/domain/chat/dto/service/response/strategy/ChatAIResponse.java
@@ -1,0 +1,5 @@
+package org.ktb.chatbotbe.domain.chat.dto.service.response.strategy;
+
+
+public interface ChatAIResponse {
+}
