+++ src/main/java/org/ktb/chatbotbe/domain/chat/dto/service/response/strategy/DoneResponse.java
@@ -1,0 +1,9 @@
+package org.ktb.chatbotbe.domain.chat.dto.service.response.strategy;
+
+import lombok.Getter;
+import org.ktb.chatbotbe.domain.chat.dto.service.response.ChatMessageType;
+
+@Getter
+public class DoneResponse implements ChatAIResponse {
+    private final ChatMessageType type = ChatMessageType.DONE;
+}
