+++ src/main/java/org/ktb/chatbotbe/domain/chat/repository/ChatCustomRepositoryImpl.java
@@ -20,6 +20,7 @@ public class ChatCustomRepositoryImpl implements ChatCustomRepository {
     public List<Chat> findAllByUser(User user) {
         BooleanBuilder builder = new BooleanBuilder();
         builder.and(chat.deletedAt.isNull());
+        builder.and(chat.user.eq(user));
         return jpaQueryFactory.selectFrom(chat)
                 .where(builder)
                 .orderBy(chat.createdAt.desc())
