+++ src/test/java/org/ktb/chatbotbe/domain/chat/service/ChatServiceTest.java
@@ -6,9 +6,9 @@
 //import org.junit.jupiter.api.Test;
 //import org.junit.jupiter.api.extension.ExtendWith;
 //import org.ktb.chatbotbe.domain.chat.dto.controller.request.ChatMessageCreateRequest;
-//import org.ktb.chatbotbe.domain.chat.dto.service.response.ChatMessageResponse;
+//import org.ktb.chatbotbe.domain.chat.dto.service.response.strategy.ChatMessageResponse;
 //import org.ktb.chatbotbe.domain.chat.dto.service.response.ChatResponse;
-//import org.ktb.chatbotbe.domain.chat.dto.service.response.NewChatResponse;
+//import org.ktb.chatbotbe.domain.chat.dto.service.response.strategy.NewChatResponse;
 //import org.ktb.chatbotbe.domain.chat.entity.Chat;
 //import org.ktb.chatbotbe.domain.chat.entity.ChatMessage;
 //import org.ktb.chatbotbe.domain.chat.exception.UnauthorizedUserChatException;
